lp-233 : extract gateway test to separate file
