+++ src/presentation/test/integration/registration/gateway.test.ts
@@ -1,0 +1,64 @@
+import request from "supertest";
+import { createApiClient } from "@companieshouse/api-sdk-node";
+import {
+  LimitedPartnershipsService,
+  NameEndingType,
+} from "@companieshouse/api-sdk-node/dist/services/limited-partnerships";
+import TransactionService from "@companieshouse/api-sdk-node/dist/services/transaction/service";
+
+import appRealDependencies from "../../../../app";
+import { appDevDependencies } from "../../../../config/dev-dependencies";
+import {
+  NAME_URL,
+  registrationRoutingName,
+} from "../../../controller/registration/Routing";
+
+jest.mock("@companieshouse/api-sdk-node");
+
+const mockCreateApiClient = createApiClient as jest.Mock;
+mockCreateApiClient.mockReturnValue({
+  transaction: {
+    ...TransactionService.prototype,
+    postTransaction: () => ({
+      httpStatusCode: 201,
+      resource: {
+        id: appDevDependencies.registrationGateway.transactionId,
+      },
+    }),
+  },
+  limitedPartnershipsService: {
+    ...LimitedPartnershipsService.prototype,
+    postLimitedPartnership: () => ({
+      httpStatusCode: 201,
+      resource: {
+        id: appDevDependencies.registrationGateway.submissionId,
+      },
+    }),
+  },
+});
+
+describe("Create transaction and the first submission", () => {
+  beforeAll(() => {
+    appDevDependencies.registrationGateway.feedLimitedPartnerships([]);
+    appDevDependencies.registrationGateway.feedErrors([]);
+  });
+
+  it("should create a transaction and the first submission", async () => {
+    const url = appDevDependencies.registrationController.insertIdsInUrl(
+      NAME_URL,
+      appDevDependencies.registrationGateway.transactionId,
+      appDevDependencies.registrationGateway.submissionId
+    );
+
+    const res = await request(appRealDependencies).post(url).send({
+      pageType: registrationRoutingName.pageType,
+      partnership_name: "Test Limited Partnership",
+      name_ending: NameEndingType.LIMITED_PARTNERSHIP,
+    });
+
+    const partialRedirectUrl = `/limited-partnerships/transaction/${appDevDependencies.registrationGateway.transactionId}/submission/${appDevDependencies.registrationGateway.submissionId}/email`;
+
+    expect(res.status).toBe(302);
+    expect(res.text).toContain(`Redirecting to ${partialRedirectUrl}`);
+  });
+});
