+++ src/views/email.njk
@@ -21,7 +21,7 @@
         },
         id: "registered-email",
         name: "registeredEmail",
-        value: props.data.email,
+        value: props.data.limitedPartnership.email,
         type: "email",
         attributes : {
           required : true
